Homework21 refactoring and adding POM
