+++ build.gradle
@@ -10,8 +10,8 @@ repositories {
 }
 
 dependencies {
-    testImplementation 'org.testng:testng:7.7.0'
-    implementation 'org.seleniumhq.selenium:selenium-java:4.11.0'
+    testImplementation 'org.testng:testng:7.6.1'
+    implementation 'org.seleniumhq.selenium:selenium-java:4.5.0'
     testImplementation 'io.github.bonigarcia:webdrivermanager:5.4.1'
 }
