+++ src/test/java/Homework16.java
@@ -1,21 +1,11 @@
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
-import org.testng.Assert;
 import org.testng.annotations.Test;
-
 public class Homework16 extends BaseTest {
 
     @Test
     public void registrationNavigation() {
-        openLoginUrl();
-        clickRegistrationBtn();
-
-        String registrationUrl = "https://qa.koel.app/registration";
-        Assert.assertEquals(driver.getCurrentUrl(), registrationUrl);
-    }
-    private void clickRegistrationBtn() {
-        WebElement registrationLink = driver.findElement(By.cssSelector("a[href='registration']"));
-        registrationLink.click();
+        basepage.openLoginUrl(url);
+        registrationPage.clickRegistrationBtn();
+        registrationPage.checkUrl();
     }
 }
