+++ src/test/java/Pages/RegistrationPage.java
@@ -1,0 +1,22 @@
+package Pages;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.testng.Assert;
+
+public class RegistrationPage extends BasePage{
+    public RegistrationPage(WebDriver driver, WebDriverWait wait, Actions actions) {
+        super(driver, wait, actions);
+    }
+    public void clickRegistrationBtn() {
+        WebElement registrationLink = driver.findElement(By.cssSelector("a[href='registration']"));
+        registrationLink.click();
+    }
+    public void checkUrl(){
+        String registrationUrl = "https://qa.koel.app/registration";
+        Assert.assertEquals(driver.getCurrentUrl(), registrationUrl);
+    }
+}
