+++ src/test/resources/SmokeTest.xml
@@ -2,7 +2,6 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
 <suite name="SmokeTest">
-    <parameter name="BaseURL" value="https://qa.koel.app/"/>
     <test name="Play a song" preserve-order="false">
         <classes>
             <class name="Homework18">
